reimplement text cursor blinking
uhoh it's looking like im using the not a starting point pr as a starting point lol
